requirements.txt
aiohttp==3.8.5
aioredis==2.0.1
aiosignal==1.3.1
annotated-types==0.5.0
anyio==3.7.1
async-timeout==4.0.2
attrs==23.1.0
charset-normalizer==3.2.0
click==8.1.6
exceptiongroup==1.1.2
fastapi==0.101.0
frozenlist==1.4.0
h11==0.14.0
idna==3.4
multidict==6.0.4
numpy==1.24.4
opencv-python-headless==4.8.0.74
Pillow==10.0.0
pydantic==2.1.1
pydantic-core==2.4.0
python-multipart==0.0.6
sniffio==1.3.0
starlette==0.27.0
typing-extensions==4.7.1
uvicorn==0.23.2
yarl==1.9.2
